+++ src/main/java/kitchenpos/ordertable/application/TableService.java
@@ -1,6 +1,6 @@
 package kitchenpos.ordertable.application;
 
-import kitchenpos.dao.OrderDao;
+import kitchenpos.order.dao.OrderDao;
 import kitchenpos.dao.OrderTableDao;
 import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.ordertable.domain.OrderTable;
