+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -14,8 +14,8 @@ import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.product.domain.Product;
-import kitchenpos.request.OrderCreateRequest;
-import kitchenpos.request.OrderLineItemDto;
+import kitchenpos.order.request.OrderCreateRequest;
+import kitchenpos.order.request.OrderLineItemDto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
